--- src/main/java/com/example/spring/domain/auth/domain/Authority.java
@@ -1,21 +1,0 @@
-package com.example.spring.domain.auth.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Authority {
-
-    @Id
-    @Column(name = "authority_name", length = 50)
-    private String authorityName;
-}
