--- src/main/java/com/example/spring/domain/member/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.example.spring.domain.member;
-
-import com.example.spring.domain.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
