+++ src/main/java/com/example/spring/domain/member/application/MemberService.java
@@ -1,6 +1,7 @@
-package com.example.spring.domain.member;
+package com.example.spring.domain.member.application;
 
 import com.example.spring.domain.member.domain.Member;
+import com.example.spring.domain.member.domain.repository.MemberRepository;
 import com.example.spring.global.apiResponse.code.status.ErrorStatus;
 import com.example.spring.global.apiResponse.exception.GeneralException;
 import lombok.RequiredArgsConstructor;
