+++ src/main/java/com/example/spring/domain/member/domain/repository/MemberRepository.java
@@ -1,10 +1,10 @@
-package com.example.spring.domain.user.domain.repository;
+package com.example.spring.domain.member.domain.repository;
 
-import com.example.spring.domain.user.domain.Member;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
+import com.example.spring.domain.member.domain.Member;
 
 import java.util.Optional;
 
@@ -12,7 +12,6 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Boolean existsByEmail(String email);
 
-    @EntityGraph(attributePaths = "authorities")
     Optional<Member> findByEmail(String email);
 
 //    Optional<Member> findMemberByUsernameAndRefreshToken(String username, String refreshToken);
