--- src/main/java/com/example/spring/domain/user/domain/gender.java
@@ -1,5 +1,0 @@
-package com.example.spring.domain.user.domain;
-
-public enum gender {
-    MALE, FEMALE
-}
